double for score

+++ uct.h
@@ -16,7 +16,7 @@ private:
 	std::vector<std::pair<libataxx::Move, uct_node *> > children;
 	std::vector<libataxx::Move>    *unvisited { nullptr };
 	uint64_t                        visited   { 0 };
-	uint64_t                        score     { 0 };
+	double                          score     { 0. };
 
 	uct_node *add_child(const libataxx::Move & m);
 	void      update_stats(const int result);
@@ -43,7 +43,7 @@ public:
 	const libataxx::Move get_causing_move() const;
 
 	const std::vector<std::pair<libataxx::Move, uct_node *> > & get_children() const;
-	void      update_stats(const uint64_t visited, const uint64_t score);
+	void      update_stats(const uint64_t visited, const double score);
 	uint64_t  get_visit_count();
-	uint64_t  get_score_count();
+	double    get_score_count();
 };
